+++ CHANGELOG.md
@@ -13,6 +13,7 @@ List of changes that are finished but not yet released in any final version.
  - [PR-354](https://github.com/Cognifide/knotx/pull/354) - Improvement - added CSRF protection and ability to configure it on any route.
  - [PR-357](https://github.com/Cognifide/knotx/pull/357) - Fix for #356 - Create AdapterProxies on serviceKnot start instead for each request
  - [PR-364](https://github.com/Cognifide/knotx/pull/364) - Fixed #358 - Use compiled regexp instead of String while matching the fragment
+ - [PR-361](https://github.com/Cognifide/knotx/pull/361) - Improved Adapter / Knots creation - we do not create them with every request.
  
 ## Version 1.1.2
  - [PR-318](https://github.com/Cognifide/knotx/pull/318) - Knot.x returns exit code `30` in case of missing config